requirements.txt
flask==2.1.0
requests>=2.25.0
pandas==1.3.0
python_dotenv
os
pymongo